Add floating button to admiral lib
