+++ lib/src/widgets/controls/floating_button/floating_button_scheme.dart
@@ -2,8 +2,8 @@ import 'package:admiralui_flutter/admiralui_flutter.dart';
 import 'package:admiralui_flutter/theme/admiral_color.dart';
 import 'package:flutter/material.dart';
 
-class ThemeSwitchWidgetScheme {
-  ThemeSwitchWidgetScheme({
+class AdmiralFloatingButtonScheme {
+  AdmiralFloatingButtonScheme({
     required this.theme,
   }) {
     font = theme.fonts.body1;
